requirements.txt
flask==3.1.*
pillow==11.*
gunicorn==23.*
